+++ CHANGELOG.md
@@ -8,14 +8,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.14] - 2024-10-24
+
+Data Contract CLI now supports the Open Data Contract Standard (ODCS) v3.0.0.
+
+### Added
+- `datacontract test` now also supports ODCS v3 data contract format
+- `datacontract export --format odcs_v3`: Export to Open Data Contract Standard v3.0.0 (#460)
+- Support for import from Iceberg table definitions.
+- Support for decimal logical type on avro export.
+- Support for custom Trino types
+
+### Changed
+- `datacontract import --format odcs`: Now supports ODSC v3.0.0 files (#474)
+- `datacontract export --format odcs`: Now creates v3.0.0 Open Data Contract Standard files (alias to odcs_v3). Old versions are still available as format `odcs_v2`. (#460)
 
 ### Fixed
 - fix timestamp serialization from parquet -> duckdb (#472)
 
-### Added
-- Support for import from Iceberg table definitions.
-- Support for decimal logical type on avro export.
-- Support for custom Trino types
 
 ## [0.10.13] - 2024-09-20